my ideas for init scene that controls evreything
My thought right now on how the persistent object/Controlling scene Start menu will open the Initialization
Scene as single The init scene can then open and close scenes as needed (town/shop/map/dialogue) What I have
just made opens up the ShopTestScene then the TestEncounterScene since it has a dialogue in it My thinking is
that once the dialogue us done, it just removes its own scene from the init scene Then the town scene can just
chill till the map/something else is clicked. the "HUD" objects like open map/journal would be in the init
object but the town scene can open up the shop scene for the desired shop as it sees fit.
Idk if this is the right approach or the one we want to go  down but it's an idea for now. I would like to hear what other people think of this... Is it overkill or do we want something that's simpler?