Reimplement encounter randomization
It has been reimplemented as a randomly shuffled queue, allowing
the player to not experience the same random encounters again
until all encounters in the queue have been interacted with.

Once all encounters have been seen, the queue will be reshuffled
again so the player sees the same encounters, but in a different
order.